requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-cors-middleware==1.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
gunicorn==20.0.4
